+++ tests/test_db.py
@@ -26,7 +26,7 @@ class DbTest(unittest.TestCase):
         version = db.version_get(conn)
         conn.close()
 
-        self.assertEqual(version, 18)
+        self.assertEqual(version, 19)
 
     @use_repository
     def test_sensor(self, config):